+++ src/Agents.jl
@@ -59,22 +59,10 @@ if display_update
             Welcome to this new major version of Agents.jl!
             Noteworthy changes:
 
-            * LightGraphs.jl dependency has been replaced by Graphs.jl.
-              No change was done to `GraphSpace`, you only need to replace
-              `using LightGraphs` with `using Graphs`.
-            * The `OpenStreetMapSpace` now uses a new depedency, LightOSM.jl,
-              which is much more performant than the previous OpenStreetMapX.jl.
-              This meas that initializing a new space is slightly different, see
-              the docstring of `OpenStreetMapSpace` for more.
-            * Agents.jl + InteractiveDynamics.jl now support native plotting for
-              open street map spaces, which is integrated in all interactive apps as well!
-            * The majority of examples of the documentation have now moved to a new repository,
-              JuliaDyamics/AgentsExampleZoo.jl. From now on, all additional examples
-              will be hosted in that repo, to keep the main documentation lean.
-            * Functions for planning and moving along routes have had their names
-              unified across Pathfinding and OpenStreetMap modules. The names now are
-              `plan_route!` and `move_along_route!` and are accessible from the top
-              level scope.
+            * Schedulers have been reworked to be more performant and allocate
+              less. This means that most scheduler names have been deprecated from
+              functions to types, such as `by_type` -> `ByType`. See changelog for
+              full list!
             * See the CHANGELOG.md or online docs for more!
             """;
             color = :light_magenta,